PRR-22 [refactor]: 위치 수정
