--- src/docs/asciidoc/auth.adoc
@@ -1,9 +1,0 @@
-[[Auth-API]]
-= *Auth API*
-
-[[로그인-API]]
-== *1. 로그인-API*
-
-=== *1-1* 성공
-
-operation::auth-controller-test/user-login-success[snippets='http-request,request-body,response-body']
