+++ src/main/kotlin/gongback/pureureum/PureureumApplication.kt
@@ -5,7 +5,7 @@ import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
 
 @SpringBootApplication
-@ConfigurationPropertiesScan(basePackages = ["gongback.pureureum.api.properties"])
+@ConfigurationPropertiesScan(basePackages = ["gongback.pureureum.application.properties"])
 class PureureumApplication
 
 fun main(args: Array<String>) {
