--- src/main/kotlin/gongback/pureureum/api/dto/AuthDtos.kt
@@ -1,9 +1,0 @@
-package gongback.pureureum.api.dto
-
-import jakarta.validation.constraints.Email
-
-data class LoginReq(
-    @field:Email
-    val email: String,
-    val password: String
-)
