--- src/main/kotlin/gongback/pureureum/api/dto/ErrorCode.kt
@@ -1,11 +1,0 @@
-package gongback.pureureum.api.dto
-
-import org.springframework.http.HttpStatus
-
-enum class ErrorCode(
-    val httpStatus: HttpStatus,
-    val message: String
-) {
-    UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "UNAUTHORIZED!"),
-    FORBIDDEN(HttpStatus.FORBIDDEN, "FORBIDDEN!")
-}
