+++ src/main/kotlin/gongback/pureureum/presentation/api/ApiResponse.kt
@@ -1,4 +1,4 @@
-package gongback.pureureum.api.dto
+package gongback.pureureum.presentation.api
 
 data class ApiResponse<T>(
     val message: String? = "",
